# apps/api/src/openapi.yaml
openapi: 3.0.3
info:
  title: Arkeo Racing API
  description: API for the Arkeo Racing game - a horse racing simulation powered by blockchain service provider metrics.
  version: 0.1.0
  contact:
    name: Arkeo Racing
servers:
  - url: /api
    description: API base path
...
paths:
...
  /races:
    get:
      tags: [Races]
      summary: List races
      description: Returns paginated list of races
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of races
          content:
            application/json:
              schema:
                type: object
                properties:
                  races:
                    type: array
                    items:
                      $ref: '#/components/schemas/RaceSummary'
                  total:
                    type: integer
  /races/{raceId}:
    get:
      tags: [Races]
      summary: Get race by ID
      parameters:
        - $ref: '#/components/parameters/raceId'
      responses:
        '200':
          description: Race details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Race'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /races/{raceId}/selection:
    get:
      tags: [Betting]
      summary: Get user's selection
      description: Returns current user's horse selection for the race
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/raceId'
      responses:
        '200':
          description: Current selection
          content:
            application/json:
              schema:
                type: object
                properties:
                  selection:
                    $ref: '#/components/schemas/Selection'
    delete:
      tags: [Betting]
      summary: Clear selection
      description: Removes the current user's horse selection for the race
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/raceId'
      responses:
        '200':
          description: Selection cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  selection:
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Betting]
      summary: Set selection
      description: Select a horse for the race
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/raceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [raceHorseId]
              properties:
                raceHorseId:
                  type: string
      responses:
        '200':
          description: Selection saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Selection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Picks closed
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: racing_session
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  parameters:
    raceId:
      name: raceId
      in: path
      required: true
      schema:
        type: string
      description: Race UUID
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: invalid_request
...
    Race:
      type: object
      properties:
        raceId:
          type: string
        status:
          type: string
          enum: [picking, starting, live, finished, voided]
        track:
          type: object
          properties:
            name:
              type: string
        weather:
          type: object
          properties:
            name:
              type: string
        pickCloseAt:
          type: string
          format: date-time
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
          nullable: true
        horses:
          type: array
          items:
            $ref: '#/components/schemas/RaceHorse'
    RaceSummary:
      type: object
      properties:
        raceId:
          type: string
        status:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
          nullable: true
    RaceHorse:
      type: object
      properties:
        raceHorseId:
          type: string
        displayName:
          type: string
        handicapTier:
          type: string
          enum: [Light, Standard, Heavy]
        formScore:
          type: number
        difficultyMultiplier:
          type: number
        archetype:
          type: string
          enum: [front_runner, stalker, stretch_runner, grinder, burst, erratic]
        temperament:
          type: string
          enum: [calm, normal, volatile]
        serviceType:
          type: object
          properties:
            displayName:
              type: string
            iconKey:
              type: string
            colorHex:
              type: string
        assignedProvider:
          type: object
          nullable: true
          properties:
            providerPubkey:
              type: string
            moniker:
              type: string
              nullable: true
        position:
          type: number
        placement:
          type: integer
          nullable: true
...
    Selection:
      type: object
      properties:
        selectionId:
          type: string
        raceHorseId:
          type: string
        createdAt:
          type: string
          format: date-time
...